Refactor upload adapter
