+++ src/javascript/CKEditor/Picker/JahiaUpload/index.js
@@ -1,0 +1,1 @@
+export {JahiaUpload} from './JahiaUpload';
