+++ src/javascript/CKEditor/Picker/JahiaUpload/uploadAdapter.gql-queries.js
